requirements.txt
requests
dnspython
jinja2
urllib3
dnspython
